Only download if the file doesn't exist (libdeflate)

+++ cos/packages/libdeflate.lua
@@ -7,16 +7,17 @@ libdeflate.metadata = {
     dependencies = {"libdeflate"}
 }
 libdeflate.startup = function()
-    local hweb, err = http.get("https://raw.githubusercontent.com/MCJack123/CC-Archive/master/LibDeflate.lua")
-    if not hweb then
-        log("Failed to download LibDeflate.lua: " .. err, false, "error")
-        error("Failed to download LibDeflate.lua: " .. err)
+    if not fs.exists("/cos/lib/LibDeflate.lua") then
+        local hweb, err = http.get("https://raw.githubusercontent.com/MCJack123/CC-Archive/master/LibDeflate.lua")
+        if not hweb then
+            log("Failed to download LibDeflate.lua: " .. err, false, "error")
+            error("Failed to download LibDeflate.lua: " .. err)
+        end
+        local hfile = fs.open("/cos/lib/LibDeflate.lua", "w")
+        hfile.write(hweb.readAll())
+        hfile.close()
+        hweb.close()
     end
-    local hfile = fs.open("/cos/lib/LibDeflate.lua", "w")
-    hfile.write(hweb.readAll())
-    hfile.close()
-    hweb.close()
-
     return true
 end
